backend/requirements.txt
fastapi
uvicorn
google-cloud-storage
pydantic
python-dotenv
